refactoring component
