+++ composer.json
@@ -62,7 +62,7 @@
         "test": "@test:sqlite",
         "test:pint": "./vendor/bin/pint --test",
         "test:sqlite": [
-            "./vendor/bin/pest --compact"
+            "./vendor/bin/pest --compact --bail"
         ],
         "test:mysql":  [
             "./vendor/bin/pest --compact --configuration phpunit.mysql.xml"
