+++ src/PowerGridComponent.php
@@ -4,21 +4,15 @@ namespace PowerComponents\LivewirePowerGrid;
 
 use Exception;
 use Illuminate\Contracts\Foundation\Application;
-use Illuminate\Contracts\Pagination\LengthAwarePaginator;
 use Illuminate\Contracts\View\{Factory, View};
 use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
 use Illuminate\Database\Eloquent\Concerns\HasAttributes;
-use Illuminate\Database\Eloquent\Relations\MorphToMany;
 use Illuminate\Database\Query\Builder as QueryBuilder;
-use Illuminate\Pagination\Paginator;
-use Illuminate\Support\{Collection as BaseCollection, Facades\Cache, Str};
+use Illuminate\Pagination\{LengthAwarePaginator, Paginator};
+use Illuminate\Support\{Collection as BaseCollection, Facades\Cache};
 use Livewire\{Component, WithPagination};
 use PowerComponents\LivewirePowerGrid\Helpers\{
-    ActionRender,
-    ActionRules,
-    Builder,
     Collection,
-    SqlSupport
 };
 use PowerComponents\LivewirePowerGrid\Themes\ThemeBase;
 use PowerComponents\LivewirePowerGrid\Traits\{HasFilter,
@@ -49,8 +43,6 @@ class PowerGridComponent extends Component
     public array $filtered = [];
 
     public string $primaryKey = 'id';
-
-    public bool $isCollection = false;
 
     public string $currentTable = '';
 
@@ -169,39 +161,6 @@ class PowerGridComponent extends Component
         $this->readyToLoad = true;
     }
 
-    /**
-     * @throws Exception|Throwable
-     */
-    public function render(): Application|Factory|View
-    {
-        /** @var ThemeBase $themeBase */
-        $themeBase = PowerGrid::theme($this->template() ?? powerGridTheme());
-
-        $this->powerGridTheme = $themeBase->apply();
-
-        $this->columns = collect($this->columns)->map(function ($column) {
-            return (object) $column;
-        })->toArray();
-
-        $this->relationSearch = $this->relationSearch();
-        $this->searchMorphs   = $this->searchMorphs();
-
-        $data = $this->getCachedData();
-
-        if (method_exists($this, 'initActions')) {
-            $this->initActions();
-
-            if (method_exists($this, 'header')) {
-                $this->headers = $this->header();
-            }
-        }
-
-        /** @phpstan-ignore-next-line */
-        $this->totalCurrentPage = method_exists($data, 'items') ? count($data->items()) : $data->count();
-
-        return $this->renderView($data);
-    }
-
     private function getCachedData(): mixed
     {
         if (!Cache::supportsTags() || !boolval(data_get($this->setUp, 'cache.enabled'))) {
@@ -245,234 +204,11 @@ class PowerGridComponent extends Component
     }
 
     /**
-     * @throws Exception
-     * @throws Throwable
-     */
-    public function fillData(): mixed
-    {
-        /** @var EloquentBuilder|QueryBuilder|BaseCollection|BaseCollection|MorphToMany $datasource */
-        $datasource = (!empty($this->datasource)) ? $this->datasource : $this->datasource();
-
-        /** @phpstan-ignore-next-line */
-        if (is_array($this->datasource())) {
-            /** @phpstan-ignore-next-line */
-            $datasource = collect($this->datasource());
-        }
-
-        $this->isCollection = is_a((object) $datasource, BaseCollection::class);
-
-        if ($this->isCollection) {
-            /** @var EloquentBuilder|QueryBuilder|BaseCollection|BaseCollection $datasource */
-            cache()->forget($this->id);
-
-            $filters = Collection::make($this->resolveCollection($datasource), $this)
-                ->filterContains()
-                ->filter();
-
-            $results = $this->applySorting($filters);
-
-            if ($this->headerTotalColumn || $this->footerTotalColumn) {
-                $this->withoutPaginatedData = $results->values()
-                    ->map(fn ($item) => (array) $item);
-            }
-
-            if ($results->count()) {
-                $this->filtered = $results->pluck($this->primaryKey)->toArray();
-
-                $paginated = Collection::paginate($results, intval(data_get($this->setUp, 'footer.perPage')));
-                $results   = $paginated->setCollection($this->transform($paginated->getCollection()));
-            }
-
-            self::resolveDetailRow($results);
-
-            return $results;
-        }
-
-        if ($datasource instanceof QueryBuilder) {
-            $this->currentTable = $datasource->from;
-        } else {
-            /** @phpstan-ignore-next-line */
-            $this->currentTable = $datasource->getModel()->getTable();
-        }
-
-        $sortField = Str::of($this->sortField)->contains('.') || $this->ignoreTablePrefix
-            ? $this->sortField : $this->currentTable . '.' . $this->sortField;
-
-        /** @var EloquentBuilder|QueryBuilder|MorphToMany $results */
-        $results = $this->resolveModel($datasource)
-            ->where(
-                fn (EloquentBuilder|QueryBuilder $query) => Builder::make($query, $this)
-                ->filterContains()
-                ->filter()
-            );
-
-        if (!$datasource instanceof QueryBuilder) {
-            /** @var EloquentBuilder|MorphToMany $softDeleteResults */
-            $softDeleteResults = $results;
-
-            $results = self::applySoftDeletes($softDeleteResults);
-        }
-
-        if ($this->multiSort) {
-            foreach ($this->sortArray as $sortField => $direction) {
-                $sortField = Str::of($sortField)->contains('.') || $this->ignoreTablePrefix ? $sortField : $this->currentTable . '.' . $sortField;
-
-                if ($this->withSortStringNumber) {
-                    $results = self::applyWithSortStringNumber($results, $sortField, $direction);
-                }
-                $results = $results->orderBy($sortField, $direction);
-            }
-        } else {
-            $results = self::applyWithSortStringNumber($results, $sortField);
-            $results = $results->orderBy($sortField, $this->sortDirection);
-        }
-
-        self::applyTotalColumn($results);
-
-        $results = self::applyPerPage($results);
-
-        self::resolveDetailRow($results);
-
-        if (method_exists($results, 'total')) {
-            $this->total = $results->total();
-        }
-
-        /** @phpstan-ignore-next-line */
-        return $results->setCollection($this->transform($results->getCollection()));
-    }
-
-    private function applyTotalColumn(EloquentBuilder|QueryBuilder|MorphToMany $results): void
-    {
-        if ($this->headerTotalColumn || $this->footerTotalColumn) {
-            $this->withoutPaginatedData = $this->transform($results->get());
-        }
-    }
-
-    /**
-     * @throws Exception
-     */
-    private function applyWithSortStringNumber(
-        EloquentBuilder|QueryBuilder|MorphToMany $results,
-        string $sortField,
-        string $multiSortDirection = null
-    ): EloquentBuilder|QueryBuilder|MorphToMany {
-        if (!$this->withSortStringNumber) {
-            return $results;
-        }
-
-        $direction = $this->sortDirection;
-
-        if ($multiSortDirection) {
-            $direction = $multiSortDirection;
-        }
-
-        $sortFieldType = SqlSupport::getSortFieldType($sortField);
-
-        if (SqlSupport::isValidSortFieldType($sortFieldType)) {
-            $results->orderByRaw(SqlSupport::sortStringAsNumber($sortField) . ' ' . $direction);
-        }
-
-        return $results;
-    }
-
-    private function applyPerPage(EloquentBuilder|QueryBuilder|MorphToMany $results): LengthAwarePaginator|Paginator
-    {
-        $perPage     = intval(data_get($this->setUp, 'footer.perPage'));
-        $recordCount = strval(data_get($this->setUp, 'footer.recordCount'));
-
-        $paginate = match ($recordCount) {
-            'min'   => 'simplePaginate',
-            default => 'paginate',
-        };
-
-        if ($perPage > 0) {
-            return $results->$paginate($perPage);
-        }
-
-        return $results->$paginate($results->count());
-    }
-
-    private function resolveDetailRow(Paginator|LengthAwarePaginator|BaseCollection $results): void
-    {
-        if (!isset($this->setUp['detail'])) {
-            return;
-        }
-
-        $collection = $results;
-
-        if (!$results instanceof BaseCollection) {
-            $collection = collect($results->items());
-        }
-
-        /** @phpstan-ignore-next-line */
-        $collection->each(function ($model) {
-            $id = strval($model->{$this->primaryKey});
-
-            data_set($this->setUp, 'detail', (array) $this->setUp['detail']);
-
-            $state = data_get($this->setUp, 'detail.state.' . $id, false);
-
-            data_set($this->setUp, 'detail.state.' . $id, $state);
-        });
-    }
-
-    /**
      * @return null
      */
     public function datasource()
     {
         return null;
-    }
-
-    /**
-     * @throws Exception
-     */
-    protected function resolveCollection(array|BaseCollection|EloquentBuilder|QueryBuilder|null $datasource = null): BaseCollection
-    {
-        if (!boolval(config('livewire-powergrid.cached_data', false))) {
-            return new BaseCollection($this->datasource());
-        }
-
-        return cache()->rememberForever($this->id, function () use ($datasource) {
-            if (is_array($datasource)) {
-                return new BaseCollection($datasource);
-            }
-
-            if (is_a((object) $datasource, BaseCollection::class)) {
-                return $datasource;
-            }
-
-            /** @var array $datasource */
-            return new BaseCollection($datasource);
-        });
-    }
-
-    protected function transform(BaseCollection $results): BaseCollection
-    {
-        return $results->map(function ($row) {
-            $addColumns = $this->addColumns();
-
-            $columns = $addColumns->columns;
-
-            $columns = collect($columns);
-
-            /** @phpstan-ignore-next-line */
-            $data = $columns->mapWithKeys(fn ($column, $columnName) => (object) [$columnName => $column((object) $row)]);
-
-            if (method_exists($this, 'actions') && count($this->actions())) {
-                $actions = resolve(ActionRender::class)->resolveActionRender($this->actions(), (object) $row);
-            }
-
-            if (count($this->actionRules())) {
-                $rules = resolve(ActionRules::class)->resolveRules($this->actionRules(), (object) $row);
-            }
-
-            $mergedData = $data->merge($rules ?? [])->merge($actions ?? []);
-
-            return $row instanceof \Illuminate\Database\Eloquent\Model
-                ? tap($row)->forceFill($mergedData->toArray())
-                : (object) $mergedData->toArray();
-        });
     }
 
     public function addColumns(): PowerGridColumns|PowerGridEloquent
@@ -483,27 +219,6 @@ class PowerGridComponent extends Component
     public function actionRules(): array
     {
         return [];
-    }
-
-    public function resolveModel(array|BaseCollection|null|MorphToMany|EloquentBuilder|QueryBuilder $datasource = null): mixed
-    {
-        if (blank($datasource)) {
-            return $this->datasource();
-        }
-
-        return $datasource;
-    }
-
-    private function renderView(mixed $data): Application|Factory|View
-    {
-        /** @phpstan-var view-string $view */
-        $view = $this->powerGridTheme->layout->table;
-
-        return view($view, [
-            'data'  => $data,
-            'theme' => $this->powerGridTheme,
-            'table' => 'livewire-powergrid::components.table',
-        ]);
     }
 
     public function checkedValues(): array
@@ -560,6 +275,21 @@ class PowerGridComponent extends Component
         $this->emitSelf('$refresh', []);
     }
 
+    public function getHasColumnFiltersProperty(): bool
+    {
+        return collect($this->columns)
+                ->filter(fn ($column) => filled($column->filters))->count() > 0;
+    }
+
+    /**
+     * @throws Exception
+     * @throws Throwable
+     */
+    public function fillData(): BaseCollection|LengthAwarePaginator|\Illuminate\Contracts\Pagination\LengthAwarePaginator|Paginator
+    {
+        return ProcessDataSourceToRender::fillData($this)->get();
+    }
+
     protected function powerGridListeners(): array
     {
         return [
@@ -573,17 +303,56 @@ class PowerGridComponent extends Component
         ];
     }
 
-    public function getHasColumnFiltersProperty(): bool
-    {
-        return collect($this->columns)
-                ->filter(fn ($column) => filled($column->filters))->count() > 0;
-    }
-
     /**
      * @return array
      */
     protected function getListeners()
     {
         return $this->powerGridListeners();
+    }
+
+    private function renderView(mixed $data): Application|Factory|View
+    {
+        /** @phpstan-var view-string $view */
+        $view = $this->powerGridTheme->layout->table;
+
+        return view($view, [
+            'data'  => $data,
+            'theme' => $this->powerGridTheme,
+            'table' => 'livewire-powergrid::components.table',
+        ]);
+    }
+
+    /**
+     * @throws Exception|Throwable
+     */
+    public function render(): Application|Factory|View
+    {
+        /** @var ThemeBase $themeBase */
+        $themeBase = PowerGrid::theme($this->template() ?? powerGridTheme());
+
+        $this->powerGridTheme = $themeBase->apply();
+
+        $this->columns = collect($this->columns)->map(function ($column) {
+            return (object) $column;
+        })->toArray();
+
+        $this->relationSearch = $this->relationSearch();
+        $this->searchMorphs   = $this->searchMorphs();
+
+        $data = $this->getCachedData();
+
+        if (method_exists($this, 'initActions')) {
+            $this->initActions();
+
+            if (method_exists($this, 'header')) {
+                $this->headers = $this->header();
+            }
+        }
+
+        /** @phpstan-ignore-next-line */
+        $this->totalCurrentPage = method_exists($data, 'items') ? count($data->items()) : $data->count();
+
+        return $this->renderView($data);
     }
 }
